inline utils and clean outputs
